Move component DV extraction from pyTACS to problem/constraint classes
pyTACS.getComponentDesignVars() read whichever design variable values
were last pushed into the assembler, but each problem/constraint
created from a pyTACS instance holds its own independent design
vector (TACSSystem.x). The pyTACS-level result was therefore
ambiguous: it depended on which problem last called setDesignVars.

Move getComponentDesignVars/getComponentDesignVarNums (and their
private helpers) onto TACSSystem, the shared base class of all
problems and constraints, so each call reads the calling object's own
design vector and returns an unambiguous snapshot of that object's
sizing state. pyTACS keeps no copy of this API. Update the example
and tests to call the methods on a problem instance instead of the
pyTACS assembler object.
