+++ tacs/system.py
@@ -6,6 +6,8 @@ pySystem
 # Imports
 # =============================================================================
 
+import numpy as np
+
 from tacs.utilities import BaseUI
 
 
@@ -141,6 +143,188 @@ class TACSSystem(BaseUI):
         xub = self.assembler.createDesignVec()
         self.assembler.getDesignVarRange(xlb, xub)
         return xlb.getArray(), xub.getArray()
+
+    def getComponentDesignVars(self):
+        """
+        Get the design variable groups and their current values for every component.
+
+        The returned dictionary is keyed by component description and each entry maps design
+        variable group names to their current values. Group names and value types match the
+        keyword arguments of the corresponding constitutive class constructor, so the values
+        can be passed directly back to the constructor inside an elemCallBack function to
+        recreate the current sizing state in another TACS execution.
+
+        One exception: the ``lp`` group of ``LamParamFullShellConstitutive`` cannot be set
+        through the constructor; restore it with ``setLaminationParameters`` after
+        construction.
+
+        All design variable groups are always returned, whether or not their entries are
+        active design variables. Active entries reflect this problem's or constraint's
+        current design variable values, i.e. the values last passed to its
+        :meth:`setDesignVars` method; inactive entries hold the values the constitutive
+        object was constructed with. Use :meth:`getComponentDesignVarNums` to determine
+        which entries are active.
+
+        This method is collective on this object's comm; the returned dictionary is
+        identical on every processor.
+
+        Returns
+        -------
+        compDVs : dict[str, dict[str, Union[float, np.ndarray]]]
+            Dictionary of design variable group values for each component. Scalar groups
+            are returned as scalars and array groups as 1D numpy arrays.
+        """
+        if self.meshLoader is None:
+            raise self._TACSError(
+                "getComponentDesignVars requires a meshLoader. Create this "
+                "problem/constraint through the pyTACS factory methods (e.g. "
+                "createStaticProblem)."
+            )
+        self._checkUniqueComponentDescripts()
+        compDescripts = self.meshLoader.getComponentDescripts()
+        nComp = self.meshLoader.getNumComponents()
+        globalDVValues = self._gatherGlobalDesignVarValues()
+        compDVs = {}
+        for compID in range(nComp):
+            descript = compDescripts[compID]
+            con = self._getComponentConstitutive(compID)
+            groupValues = {}
+            if con is not None:
+                self._checkDVGroupsImplemented(compID, con)
+                groupValues = con.getDesignVarGroups()
+                groupNums = con.getDesignVarGroupDVNums()
+                # Overlay this object's current values onto the active entries
+                for name, nums in groupNums.items():
+                    if isinstance(nums, np.ndarray):
+                        for ii in range(len(nums)):
+                            if nums[ii] >= 0:
+                                groupValues[name][ii] = globalDVValues[nums[ii]]
+                    elif nums >= 0:
+                        groupValues[name] = globalDVValues[nums].item()
+            compDVs[descript] = groupValues
+        return compDVs
+
+    def getComponentDesignVarNums(self):
+        """
+        Get the global design variable numbers of every design variable group for every
+        component.
+
+        The returned dictionary has the same structure as the one returned by
+        :meth:`getComponentDesignVars`, but contains global design variable numbers instead
+        of values. Entries that are not active design variables have a design variable
+        number of -1. Note that design variable numbers are specific to this TACS execution
+        and should not be stored across executions.
+
+        Design variable numbers are shared by all problems and constraints within one TACS
+        execution, so this dictionary does not depend on which problem or constraint it is
+        called on.
+
+        Returns
+        -------
+        compDVNums : dict[str, dict[str, Union[int, np.ndarray]]]
+            Dictionary of design variable group numbers for each component. Scalar groups
+            are returned as integers and array groups as 1D numpy integer arrays.
+        """
+        if self.meshLoader is None:
+            raise self._TACSError(
+                "getComponentDesignVarNums requires a meshLoader. Create this "
+                "problem/constraint through the pyTACS factory methods (e.g. "
+                "createStaticProblem)."
+            )
+        self._checkUniqueComponentDescripts()
+        compDescripts = self.meshLoader.getComponentDescripts()
+        nComp = self.meshLoader.getNumComponents()
+        compDVNums = {}
+        for compID in range(nComp):
+            descript = compDescripts[compID]
+            con = self._getComponentConstitutive(compID)
+            groupNums = {}
+            if con is not None:
+                self._checkDVGroupsImplemented(compID, con)
+                groupNums = con.getDesignVarGroupDVNums()
+            compDVNums[descript] = groupNums
+        return compDVNums
+
+    def _checkUniqueComponentDescripts(self):
+        """Raise an error if any two components share a description, since the component
+        design variable dictionaries are keyed by description.
+        """
+        compDescripts = self.meshLoader.getComponentDescripts()
+        seen = set()
+        duplicates = set()
+        for descript in compDescripts:
+            if descript in seen:
+                duplicates.add(descript)
+            seen.add(descript)
+        if len(duplicates) > 0:
+            raise self._TACSError(
+                "Component descriptions must be unique to use getComponentDesignVars/"
+                "getComponentDesignVarNums, but the following descriptions are shared by "
+                f"multiple components: {sorted(duplicates)}. Rename the property groups in "
+                "the BDF file so that every component has a unique description."
+            )
+
+    def _getComponentConstitutive(self, compID):
+        """Get the single constitutive object shared by all element objects of a component.
+
+        Returns None if the component has no constitutive object. Raises an error if the
+        component's element objects hold more than one distinct constitutive object, since
+        the component design variable dictionaries cannot represent that faithfully.
+        """
+        compDescripts = self.meshLoader.getComponentDescripts()
+        conObj = None
+        numObjs = len(self.meshLoader.getElementObjectNumsForComp(compID))
+        for objIndex in range(numObjs):
+            elemObj = self.meshLoader.getElementObject(compID, objIndex)
+            elemCon = elemObj.getConstitutive()
+            if elemCon is None:
+                continue
+            if conObj is None:
+                conObj = elemCon
+            elif elemCon is not conObj:
+                raise self._TACSError(
+                    f"Component {compID} ('{compDescripts[compID]}') uses more than "
+                    "one distinct constitutive object across its element types. "
+                    "getComponentDesignVars/getComponentDesignVarNums require all element "
+                    "types in a component to share a single constitutive object."
+                )
+        return conObj
+
+    def _gatherGlobalDesignVarValues(self):
+        """Gather this problem's/constraint's current design variable values into a
+        rank-identical numpy array indexed by global design variable number.
+
+        The design vector is distributed in contiguous blocks of design variable numbers,
+        so concatenating the local arrays in rank order yields the global array.
+        """
+        localValues = self.x.getArray().copy()
+        return np.concatenate(self.comm.allgather(localValues))
+
+    def _checkDVGroupsImplemented(self, compID, con):
+        """Warn if a constitutive object's design variable groups do not cover all of its
+        design variables, since the missing DVs will be absent from the output.
+        """
+        compDescripts = self.meshLoader.getComponentDescripts()
+        numDVs = len(con.getDesignVarNums())
+        if con.getNumDesignVarGroups() == 0 and numDVs > 0:
+            self._TACSWarning(
+                f"Constitutive class '{type(con).__name__}' used by component {compID} "
+                f"('{compDescripts[compID]}') has design variables but does not "
+                "implement the design variable group interface. Its design variables will "
+                "be missing from the component design variable dictionaries."
+            )
+        else:
+            numActiveGroupEntries = 0
+            for nums in con.getDesignVarGroupDVNums().values():
+                numActiveGroupEntries += int(np.count_nonzero(np.atleast_1d(nums) >= 0))
+            if numActiveGroupEntries != numDVs:
+                self._TACSWarning(
+                    f"Constitutive class '{type(con).__name__}' used by component {compID} "
+                    f"('{compDescripts[compID]}') has {numDVs} design variables but its "
+                    f"design variable groups only cover {numActiveGroupEntries} of them. The "
+                    "remaining design variables will be missing from the component design "
+                    "variable dictionaries."
+                )
 
     def _arrayToDesignVec(self, dvArray):
         """
